Shift+Gameboy button (de-)presses key forever
Holding shift while pressing a Gameboy button on a frame will toggle
that button and also set the toggled button state for all future frames
from that point on. This means we can type lower-case A to shoot once,
but type Shift+R to start running right and keep running until we
toggle Shift+R in some future frame.