Rename bits.go functions to clarify bit handling
In the original fork of Goboy, bits used to be its own package, so
functions like bits.Reset would be descriptive. We moved everything up
to the main package so bits.Reset became just Reset. Now we renamed it
to ResetBit to again have a descriptive name for it. The same goes for
the other functions in bits.go.